backend/requirements.txt
# Qualified Health Backend Dependencies

# Web Framework
fastapi
uvicorn[standard]

# API & Validation
pydantic
pydantic-settings

# HTTP & Async
httpx
aiofiles

# AI/LLM Integration
openai
tiktoken

# Logging & Observability
structlog
python-json-logger

# Security
python-multipart
python-jose[cryptography]

# Environment
python-dotenv

# Database
python-arango

# RAG Pipeline
numpy
scipy

# Development
pytest
pytest-asyncio
black
ruff
